implemented velocity ramping for turning and moving; made path planning more modular
